+++ Vue/Cours/cListeCours.php
@@ -61,7 +61,7 @@ else{
                                 <i class="fas fa-eye h-4 w-4"></i>
                                 <span class="ml-2 text-white text-xs font-medium">Élève</span>
                             </a>
-                            <a class="flex items-center bg-red-500 h-10 px-3 rounded hover:bg-gray-300" href="index.php?uc=seance&action=supprimer&idSeance=<?= $cours->getIdSeance() ?>">
+                            <a class="flex items-center bg-red-500 h-10 px-3 rounded hover:bg-gray-300" href="index.php?uc=seance&action=supprimer&idSeance=<?= $cours->getIdSeance() ?>"  onclick="confirmDeletion(event)">
                                 <i class="fas fa-trash-alt h-4 w-4"></i>
                                 <span class="ml-2 text-white text-xs font-medium">Supprimer</span>
                             </a>
@@ -71,5 +71,10 @@ else{
             <?php } ?>
         </div>
     </div>
+    <script>
+        function confirmDeletion() {
+            return confirm('Êtes-vous sûr de vouloir supprimer l\'utilisateur ?');
+        }
+    </script>
 
     <?php include("Vue/footer.php"); ?>
